+++ src/stl_type_cast.cpp
@@ -22,7 +22,7 @@ namespace pybind
 
                 if( pybind::string_check( _obj ) == true )
                 {
-                    size_t size;
+                    uint32_t size;
                     const std::string::value_type * string_char = pybind::string_to_char_and_size( _obj, size );
 
                     if( string_char == 0 )
@@ -62,7 +62,7 @@ namespace pybind
 
                 if( pybind::unicode_check( _obj ) == true )
                 {
-                    size_t size = 0;
+                    uint32_t size = 0;
                     const wchar_t * value_char = pybind::unicode_to_wchar_and_size( _obj, size );
 
                     if( value_char == nullptr )
@@ -102,7 +102,7 @@ namespace pybind
 
                 if( pybind::string_check( _obj ) == true )
                 {
-                    size_t size;
+                    uint32_t size;
                     const stdex::string::value_type * string_char = pybind::string_to_char_and_size( _obj, size );
 
                     if( string_char == 0 )
@@ -141,7 +141,7 @@ namespace pybind
 
                 if( pybind::unicode_check( _obj ) == true )
                 {
-                    size_t size = 0;
+                    uint32_t size = 0;
                     const wchar_t * value_char = pybind::unicode_to_wchar_and_size( _obj, size );
 
                     if( value_char == nullptr )
